First steps with contracts (#2)
* contracts ideas

Added a couple of contracts to learn how this repo works for compiling and deploying, as well as to start thinking about ERC1155 and the general mechanics of an auction contract.

`Auction.sol` is a contract which handles the owner, duration, bidding, winning and ability to withdrawl after an auction. I tested the logic locally and it seems ok. Its not in the right ERC1155 spec.. I was just seeing how I could define an auction.

`GameItems.sol` is taken from a hello world ERC1155 contract from here:
https://docs.openzeppelin.com/contracts/4.x/erc1155

uses the helper interface/module from OpenZeppelin to get all the functionality. I haven't done much with it other than compile and deploy it.

Some good resources I've found:
https://docs.openzeppelin.com/contracts/4.x/api/token/erc1155
https://ethereum.org/en/developers/docs/standards/tokens/erc-1155/

another reference implementation:
https://github.com/enjin/erc-1155/blob/master/contracts/ERC1155.sol

Doesn't matter if we merge this or not, but I wanted to share.

* adding some yarn/node files to gitignore

* deploy file for the GameItems contract

* many auction/marketplace contracts, some deploy and interaction scripts

* found this on youtube -- good ideas here

* fixing imports on the auction just found on youtube

* updating README with some env var stuff